Remove slow from dataset test

+++ test/data/test_builtin_datasets.py
@@ -9,14 +9,14 @@ from ..common.torchtext_test_case import TorchtextTestCase
 
 
 def conditional_remove(f):
-    if os.path.isfile(f):
-        os.remove(f)
-    elif os.path.isdir(f):
-        shutil.rmtree(f)
+    if os.environ.get("TRAVIS") == "true":
+        if os.path.isfile(f):
+            os.remove(f)
+        elif os.path.isdir(f):
+            shutil.rmtree(f)
 
 
 class TestDataset(TorchtextTestCase):
-    @slow
     def test_wikitext2_legacy(self):
         from torchtext.datasets import WikiText2
         # smoke test to ensure wikitext2 works properly
@@ -52,7 +52,6 @@ class TestDataset(TorchtextTestCase):
         datafile = os.path.join(self.project_root, ".data", "wikitext-2-v1.zip")
         conditional_remove(datafile)
 
-    @slow
     def test_penntreebank_legacy(self):
         from torchtext.datasets import PennTreebank
         # smoke test to ensure penn treebank works properly
@@ -110,7 +109,6 @@ class TestDataset(TorchtextTestCase):
         datafile = os.path.join(self.project_root, ".data", "ag_news_csv.tar.gz")
         conditional_remove(datafile)
 
-    @slow
     def test_imdb(self):
         from torchtext.experimental.datasets import IMDB
         from torchtext.vocab import Vocab
